requirements.txt
scratchcommunication
scratchattach
python_dotenv
cachetools
mysql